Condition-reading refactor
JavaScript: the MegaloCondition read method now returns a success bool. Fixed the loading of references to global timers.

C++: I've now fully ported the JavaScript condition-reading code. It's significantly cleaner than the old results and produces consistent (and apparently correct) results.

Fixed Megalo::SmartFlags' bitlength getters.

+++ native/src/ReachVariantTool/game_variants/megalo_ex/enums.cpp
@@ -1,5 +1,4 @@
 #include "enums.h"
-#include <string>
 #include "../../helpers/strings.h"
 
 namespace Megalo {
@@ -12,7 +11,7 @@ namespace Megalo {
    }
    void SmartFlags::to_string(std::string& out, uint32_t value) const noexcept {
       out.clear();
-      int i = 0;
+      size_t i = 0;
       for (; i < this->count; i++) {
          if (value & (1 << i)) {
             if (!out.empty())
